add new command: `deploy`, connect with railway env vars

+++ internal/pipes/token/telegram/token.go
@@ -5,12 +5,16 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"os/exec"
+	"runtime"
 	"strings"
 
 	"github.com/abdfnx/botway/constants"
 	token_shared "github.com/abdfnx/botway/internal/pipes/token"
+	"github.com/abdfnx/looker"
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
+	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
 )
 
@@ -23,12 +27,19 @@ type model struct {
 func (m model) AddToken() {
 	botwayConfig, err := ioutil.ReadFile(token_shared.BotwayConfigPath)
 	// token, _ := token_shared.EncryptTokens(m.inputs[0].Value(), "")
+	bwPath, bwErr := looker.LookPath("botway")
+
+	if bwErr != nil {
+		fmt.Print(constants.FAIL_BACKGROUND.Render("ERROR"))
+		panic(constants.FAIL_FOREGROUND.Render(" botway is not installed\n"))
+	}
 
 	if err != nil {
 		panic(err)
 	}
 
 	tokenContent, _ := sjson.Set(string(botwayConfig), "botway.bots." + m.botName + ".bot_token", m.inputs[0].Value())
+	botPath := gjson.Get(tokenContent, "botway.bots." + m.botName + ".path").String()
 
 	remove := os.Remove(token_shared.BotwayConfigPath)
 
@@ -45,6 +56,23 @@ func (m model) AddToken() {
 	fmt.Print(constants.SUCCESS_BACKGROUND.Render("SUCCESS"))
 	fmt.Println(constants.SUCCESS_FOREGROUND.Render(" " + m.botName + " Telegram token is added successfully"))
 	// fmt.Println("Your Secret key -> " + token_shared.BoldStyle.Render(token_shared.UserSecret) + " Keep it in a safe place")
+
+	set := fmt.Sprintf("%s vars set TELEGRAM_TOKEN=%s", bwPath, m.inputs[0].Value())
+	cmd := exec.Command("bash", "-c", set)
+
+	if runtime.GOOS == "windows" {
+		cmd = exec.Command("powershell.exe", set)
+	}
+
+	cmd.Dir = botPath
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	err = cmd.Run()
+
+	if err != nil {
+		log.Printf("error: %v\n", err)
+	}
 }
 
 func initialModel(botName string) model {
